Add constraints to quantity and amount values
Quantity and amount values are capped to a maximum of 2 decimal places
and cannot go beyond the double range in Java. As such, the regex to
check for a valid format of the values is modified.

Test cases are also added to StringUtilTest and ParserUtilTest to check
for valid values.